Add files via upload
Weather App
Overview
Weather-app is a Vanilla JavaScript solo project that displays real-time weather information for a user's specified location using an API. The app allows users to search for weather conditions in different cities, and view the temperature, humidity, wind speed, and other details.
Live version
GitHub deploy
Netlify deploy
Features
Displays real-time weather information
Allows users to search for weather conditions in different cities
Shows temperature, humidity, wind speed, and other details
API Reference
Here is a link to the OpenWeatherMap API which was used to make API calls.
Usage
Once the app is running, simply enter the name of the city you want to check the weather for in the search bar and press enter. The app will then display the current weather conditions for that city.